test(configurator): Refactor tests to remove dupes
